+++ scirisweb/sc_tasks.py
@@ -28,6 +28,85 @@ celery_instance = None # Celery instance.
 
 
 ################################################################################
+### Classes
+################################################################################
+
+__all__ += ['Task']
+
+
+class Task(sc.prettyobj):
+    '''
+    A Sciris record for an asynchronous task.
+    
+    Attributes:
+        task_id (str)         -- the ID / name for the task that typically is chosen by the client
+        status (str)          -- the status of the task:
+                                    'unknown': unknown status, for example, just initialized
+                                    'error':   task failed with an actual error
+        error_text (str)      -- string giving an idea of what error has transpired
+        func_name (str)       -- string of the function name for what's called
+        args (list)           -- list containing args for the function
+        kwargs (dict)         -- dict containing kwargs for the function
+        result_id (str)       -- string for the Redis ID of the AsyncResult
+        queue_time (datetime) -- the time the task was queued for Celery
+        start_time (datetime) -- the time the task was actually started
+        stop_time (datetime)  -- the time the task completed
+        pending_time (int)    -- the time the process has been waiting to be executed on the server in seconds        
+        execution_time (int)  -- the time the process required to complete
+    '''
+    
+    def  __init__(self, task_id):
+        self.task_id        = task_id # Set the task ID (what the client typically knows as the task).
+        self.uid            = task_id # Make it the same as the task ID...WARNING, fix
+        self.status         = 'unknown' # Start the status at 'unknown'.
+        self.error_text     = None # Start the error_text at None.
+        self.func_name      = None # Start the func_name at None.
+        self.args           = None # Start the args and kwargs at None.
+        self.kwargs         = None
+        self.result_id      = None # Start with no result_id.
+        self.queue_time     = None # Start the queue, start, and stop times at None.
+        self.start_time     = None
+        self.stop_time      = None
+        self.pending_time   = None # Start the pending and execution times at None.
+        self.execution_time = None
+        return None
+    
+    def show(self):
+        print('-----------------------------')
+        print('        Task ID: %s'   % self.task_id)
+        print('         Status: %s'   % self.status)
+        print('     Error text: %s'   % self.error_text)
+        print('  Function name: %s'   % self.func_name)
+        print('  Function args: %s'   % self.args)
+        print('Function kwargs: %s'   % self.kwargs)        
+        print('      Result ID: %s'   % self.result_id)
+        print('     Queue time: %s'   % self.queue_time)        
+        print('     Start time: %s'   % self.start_time)
+        print('      Stop time: %s'   % self.stop_time)
+        print('   Pending time: %s s' % self.pending_time)        
+        print(' Execution time: %s s' % self.execution_time)  
+        print('-----------------------------')
+    
+    def jsonify(self):
+        output = {'UID':           self.uid,                    
+                  'taskId':        self.task_id,
+                  'status':        self.status,
+                  'errorText':     self.error_text,
+                  'funcName':      self.func_name,
+                  'funcArgs':      self.args,
+                  'funcKwargs':    self.kwargs,
+                  'resultId':      self.result_id,
+                  'queueTime':     self.queue_time,                
+                  'startTime':     self.start_time,
+                  'stopTime':      self.stop_time,
+                  'pendingTime':   self.pending_time,
+                  'executionTime': self.execution_time                
+        }
+        return output
+
+
+
+################################################################################
 ### Functions
 ################################################################################
